LPD-58926 Move focusInvalidElement to common utils
